+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Close crossnode repl on server destroy
 - Don't reset pvp state like lastWinPartyId or points after pvp end
 - Fix remote crash when referencing entities in vars in steps
+- Remove player collType getting set to NONE when in cutscene
 
 ## [0.7.9] 2026-05-22